refactor: do not use ServiceObject
